convert use-price-contraints to ts
